refactor: standardize naming
