+++ frontend/src/apis/CollaborationApi.ts
@@ -1,4 +1,4 @@
-import { SessionResponse } from '../types/Collaboration';
+import { SessionResponse } from '../types/CollaborationType';
 import { api } from './ApiClient';
 
 export const checkSession = async (): Promise<SessionResponse> => {
