+++ frontend/src/types/AuthType.ts
