+++ frontend/src/types/CollaborationType.ts
